Rename src/components/structure_editor to src/components/structureEditor
