+++ app/packs/src/components/structureEditor/StructureEditorContent.js
