+++ app/packs/src/components/structureEditor/StructureEditorMap.js
