+++ app/packs/src/components/structureEditor/StructureEditorModal.js
@@ -12,11 +12,11 @@ import Select from 'react-select';
 import UserStore from 'src/stores/alt/stores/UserStore';
 import UIStore from 'src/stores/alt/stores/UIStore';
 import StructureEditor from 'src/models/StructureEditor';
-import EditorAttrs from 'src/components/structure_editor/StructureEditorSet';
-import ChemDrawEditor from 'src/components/structure_editor/ChemDrawEditor';
-import MarvinjsEditor from 'src/components/structure_editor/MarvinjsEditor';
-import KetcherEditor from 'src/components/structure_editor/KetcherEditor';
-import loadScripts from 'src/components/structure_editor/loadScripts';
+import EditorAttrs from 'src/components/structureEditor/StructureEditorSet';
+import ChemDrawEditor from 'src/components/structureEditor/ChemDrawEditor';
+import MarvinjsEditor from 'src/components/structureEditor/MarvinjsEditor';
+import KetcherEditor from 'src/components/structureEditor/KetcherEditor';
+import loadScripts from 'src/components/structureEditor/loadScripts';
 
 const EditorList = (props) => {
   const { options, fnChange, value } = props;
