+++ app/packs/src/components/structureEditor/StructureEditorSet.js
