+++ app/packs/src/components/structureEditor/loadScripts.js
